Refactored test files
